Bot and Scrapper connected

+++ bot/src/main/java/edu/java/bot/client/ScrapperClient.java
@@ -4,7 +4,6 @@ import edu.java.bot.controller.dto.AddLinkRequest;
 import edu.java.bot.controller.dto.LinkResponse;
 import edu.java.bot.controller.dto.ListLinkResponse;
 import edu.java.bot.controller.dto.RemoveLinkRequest;
-import io.swagger.v3.oas.models.security.SecurityScheme;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
@@ -12,25 +11,26 @@ import org.springframework.web.service.annotation.DeleteExchange;
 import org.springframework.web.service.annotation.GetExchange;
 import org.springframework.web.service.annotation.PostExchange;
 
+
 public interface ScrapperClient {
     @PostExchange("/tg-chat/{id}")
-    void createTgChat(@PathVariable Integer id);
+    void createTgChat(@PathVariable Long id);
 
     @DeleteExchange("/tg-chat/{id}")
-    void removeTgChat(@PathVariable Integer id);
+    void removeTgChat(@PathVariable Long id);
 
     @GetExchange("/links")
-    ListLinkResponse getLinks(@RequestHeader("Tg-Chat-Id") Integer chatId);
+    ListLinkResponse getLinks(@RequestHeader("Tg-Chat-Id") Long chatId);
 
     @PostExchange("/links")
     LinkResponse createLink(
-        @RequestHeader("Tg-Chat-Id") Integer chatId,
+        @RequestHeader("Tg-Chat-Id") Long chatId,
         @RequestBody AddLinkRequest addLinkRequest
     );
 
     @DeleteExchange("/links")
     LinkResponse deleteLink(
-        @RequestHeader("Tg-Chat-Id") Integer chatId,
+        @RequestHeader("Tg-Chat-Id") Long chatId,
         @RequestBody RemoveLinkRequest removeLinkRequest
     );
 }
